requirements.txt
requests
beautifulsoup4
python-dotenv
streamlit
pandas
xlsxwriter
tiktoken
langchain
langchain-community
huggingface-hub
faiss-cpu
unstructured
python-pptx
loguru
konlpy
langdetect
nltk
sentence-transformers